requirements.txt
flask
opencv-python
numpy
torch
torchvision
python-telegram-bot
gunicorn
